Add abstraction for GetMetricsData processing (#1325)
