+++ pkg/job/getmetricdata/compact_test.go
@@ -1,4 +1,4 @@
-package job
+package getmetricdata
 
 import (
 	"testing"
